fix error on plugin validation

+++ src/batch/transformations/connectExclusiveList.ts
@@ -22,13 +22,15 @@ export async function connectExclusiveList(
     return null;
   }
   // Because this is an exclusive list, delete records that are no longer relevant
-  const deleted = currentData[fieldKey]
-    .map((d: any) => d.id)
-    .filter((n: string) => !inputData[fieldKey + '_ids'].includes(n));
-  logger.debug('Deletions? %j', deleted);
-  if (deleted.length > 0) {
-    const batchService = new BatchService(keystone);
-    await batchService.removeAll(transformInfo.list, deleted);
+  if (currentData != null) {
+    const deleted = currentData[fieldKey]
+      .map((d: any) => d.id)
+      .filter((n: string) => !inputData[fieldKey + '_ids'].includes(n));
+    logger.debug('Deletions? %j', deleted);
+    if (deleted.length > 0) {
+      const batchService = new BatchService(keystone);
+      await batchService.removeAll(transformInfo.list, deleted);
+    }
   }
   return {
     disconnectAll: true,
